+++ src/main/java/polis/bot/Bot.java
@@ -12,11 +12,11 @@ import org.telegram.telegrambots.meta.api.objects.Message;
 import org.telegram.telegrambots.meta.api.objects.Update;
 import org.telegram.telegrambots.meta.api.objects.User;
 import org.telegram.telegrambots.meta.exceptions.TelegramApiException;
-import polis.authorization.AuthData;
 import polis.commands.NonCommand;
 import polis.commands.OkAuthCommand;
 import polis.commands.StartCommand;
 import polis.commands.SyncCommand;
+import polis.util.AuthData;
 import polis.util.IState;
 import polis.util.State;
 import polis.util.Substate;
