+++ src/main/java/polis/commands/OkAuthCommand.java
@@ -5,7 +5,7 @@ import org.slf4j.LoggerFactory;
 import org.telegram.telegrambots.meta.api.objects.Chat;
 import org.telegram.telegrambots.meta.api.objects.User;
 import org.telegram.telegrambots.meta.bots.AbsSender;
-import polis.authorization.OkAuthorization;
+import polis.ok.OkAuthorization;
 
 import java.net.URISyntaxException;
 import java.util.Properties;
