+++ src/main/java/polis/ok/OKDataCheck.java
@@ -5,8 +5,7 @@ import org.json.JSONArray;
 import org.json.JSONObject;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import polis.authorization.AuthData;
-import polis.authorization.OkAuthorization;
+import polis.util.AuthData;
 import polis.util.IState;
 import polis.util.SocialMedia;
 import polis.util.State;
